Value can have spikes and dips(jumps)
Jumps are when the currency fluctuates by a percentage of its value. it only jumps when the threshold is within bounds.

The currency will now act stable when the threshold is out of bounds and jumpy when it is within bounds. this is because the threshold sticks around the bounds.

spikes now set the value to 50.