percent_factor now uses the Gaussian equation. dosent crash easily when low
the gaussian equation will let the value fluctuate much more near thresholds of 50, but much less as the threshold deviates from it. also the bounds used by bounds_factor have been set back to [35, 65].

A temporary fix while we enter the last stages of currency testing, as the currency gets smaller, it stops fluctuating by the Vmax_mag and only by percent. debating whether this should be added or not as it adds lots of complexity