Mostly finished tests. stress-tested user trades.
big exploit in LUNA bot was exploiting the volatility of the currency. im looking at you kuki.
now instead of a perfectly linear increase/decrease from purchases, the volume of your purchase is calculated as a percentage of the market cap. the market cap being an extremely high value based on the value of each share * the total number of shares. the number of shares is also absurdly high and gradually increases, so being able to significantly manipulate the value is nearly impossible.