halfway done reorganizing the code. trying to implement trades to be used by the bot
remade the constants to be based on the tests. still need to test those constants though. the constants were based off of rough calculations and educated guesses.

cost.py incorporates taxes. the actual taxes are not negligible but are not too large either.

working on adding methods that trades will be used. needs to be in a way that makes sense and is organized/efficient.

planning on reworking the user class the same way as cryptocurrency. refactoring attributes to be attributes and not dictionaries.