added methods for a buy function. modified the formula for calculating value/cost for trade
because of a math error, the value of the currency was expected to fluctuate a lot less than it actually was, and the compound price to become negligible. this has been fixed.

almost all methods for a trade are done now.

total_shares has been increased to allow currencies to be more stable.