implemented a max_value, finished buy() method
when the market cap reaches its maximum, it wont increase, but there was not anything preventing the volume from increasing further, which would cause problems. to combat this, volume cannot surpass max_value()

the buy() method works well and should pass all tests. might test more tmrw, but for now it works just fine.

setup the sell method, but have yet to complete or test it