bot is nearly ready. can clear db, cooldown error works, can add, and simulate currencies every min
we can clear the database(crypto). deleting users might be in 1.1 idk
other than that, changing constant values is important. general control commands are helpful.

cooldown errors now work and tell you how many seconds it will take to refresh

tasks now work. so we can add and simulatr currencies with ease.

other than that, nitpicked a few bugs like not caching properly(was appending the currency to the end when caching and messing up any iterations through crypto_cache) and not decrementing the num_holdings in the user class when verifying holdings.

next, we need to add in the help commands which will act as documentation and also ensure we can change constants and maybe users